-mvvm pattern changed

+++ app/src/main/java/com/savemykeys/views/fragments/SignUpFragment.kt
@@ -8,19 +8,19 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.savemykeys.R
 import com.savemykeys.utils.AppUtils
 import com.savemykeys.viewmodel.LoginSignUpViewModel
 import com.savemykeys.viewmodel.RecordViewModel
 import com.savemykeys.views.activities.HomeActivity
-import com.savemykeys.views.listeners.SignUpViewListener
 import kotlinx.android.synthetic.main.fragment_sign_up.*
 
 /**
  * A simple [Fragment] subclass.
  */
-class SignUpFragment : Fragment(), View.OnClickListener, SignUpViewListener {
+class SignUpFragment : Fragment(), View.OnClickListener {
 
     private val TAG = "SignUpFragment"
     private lateinit var loginSignUpViewModel: LoginSignUpViewModel
@@ -39,8 +39,13 @@ class SignUpFragment : Fragment(), View.OnClickListener, SignUpViewListener {
         Log.d(TAG, "onActivityCreated()")
         loginSignUpViewModel = ViewModelProviders.of(this).get(LoginSignUpViewModel::class.java)
         recordViewModel = ViewModelProviders.of(this).get(RecordViewModel::class.java)
-        loginSignUpViewModel.setSignUpViewListener(this)
         btnSignUp.setOnClickListener(this)
+        loginSignUpViewModel.getSignUpStatus()
+            .observe(this,
+                Observer<Int> { message ->
+                    context?.let { signUpStatus(message) }
+                }
+            )
     }
 
     override fun onClick(view: View?) {
@@ -52,24 +57,22 @@ class SignUpFragment : Fragment(), View.OnClickListener, SignUpViewListener {
         }
     }
 
-    override fun showPinDoesNotMatchError(pinDoesNotMatch: Int) {
-        Log.d(TAG, "showPinDoesNotMatchError() pinDoesNotMatch")
-        activity?.let { AppUtils.showToastMessageById(it, pinDoesNotMatch) }
+    private fun signUpStatus(message: Int) {
+        Log.d(TAG, "signUpStatus() ${getString(message)}")
+        activity?.let { AppUtils.showToastMessageById(it, message) }
+        if (resources.getString(R.string.signUpSuccess).equals(
+                getString(message),
+                ignoreCase = true
+            )
+        ) {
+            recordViewModel.deleteAllRecord()
+            val intent = Intent(activity, HomeActivity::class.java)
+            startActivity(intent)
+            activity?.finish()
+        }
+
     }
 
-    override fun signUpSuccess(signUpSuccessMessage: Int) {
-        Log.d(TAG, "signUpSuccess() signUpSuccessMessage")
-        activity?.let { AppUtils.showToastMessageById(it, signUpSuccessMessage) }
-        recordViewModel.deleteAllRecord()
-        val intent = Intent(activity, HomeActivity::class.java)
-        startActivity(intent)
-        activity?.finish()
-    }
-
-    override fun showEmptyPinError(emptyPinMessage: Int) {
-        Log.d(TAG, "showEmptyPinError() emptyPinMessage")
-        activity?.let { AppUtils.showToastMessageById(it, emptyPinMessage) }
-    }
 }
